feat: header component 구현
- header index 파일 추가
- addEvent 함수 구현
- navigation 컴포넌트 추가

- #24
